Move template module outside handlers package
