+++ forgot_password/handlers/forgot_password.py
@@ -24,7 +24,7 @@ from skygear.models import Record, RecordID
 from skygear.utils.context import current_context
 from skygear.utils.db import conn
 
-from .template import FileTemplate
+from ..template import FileTemplate
 from .template_mail import TemplateMailSender
 from .util import user as user_util
